--- a/.github/change-requests/CHANGE_REQUEST_3_MCP_HTTP.md
+++ b/.github/change-requests/CHANGE_REQUEST_3_MCP_HTTP.md
@@ -1,4 +1,4 @@
-# Change Request: Convert MCP Server to HTTP-Streamable Architecture
+# Change Request 3: Convert MCP Server to HTTP-Streamable Architecture
 
 ## Hi GitHub Copilot! 👋
 
@@ -437,7 +437,9 @@ curl http://localhost:3000/health
 ---
 
 *Last Updated: September 17, 2025*  
-*Document Version: 2.0 - Restructured for Clear Implementation*
+*Document Version: 2.0 - Restructured for Clear Implementation*  
+*Branch: HTTP-MCP*  
+*Status: Ready for Implementation*
 
 #### Core Services Configuration
 ```yaml