… .github/change-requests/CHANGE_REQUEST_TRACKING.md
… .github/prompts/change-request.prompt.md
… CHANGE_REQUEST_MCP_HTTP.md
… README.md
… package.json
… simple_whisper_test.py